temp_dashboard/maee-demo-repo/requirements.txt
torch
numpy
matplotlib
scikit-learn
seaborn
gitpython
plotly
networkx
scipy 